InfoPanel and Highlight Fix
(1) dot for each species has not a color pie chart if different colors are
selected.

(2) openining the infoPanel by double clicking on an now highlight the
individual atoms rows. also remove a timer for auto deselect that
somehow made it in there. selection is now simply removed by clicking
somewhere in the background